Tidied up code
Put cell classes into a separate file. Restructured code. Put level
generation in a separate level class. Moved movement detection to root
HotrodGame.
